Adding exceptions

+++ Lab02/src/StringWorker.java
@@ -1,10 +1,16 @@
 public class StringWorker {
     public static void main(String[] args){
-        StringBuilder stringBuilder = new StringBuilder(args[0]);
-        BracketsWorker bracketsWorker = new BracketsWorker(stringBuilder);
-        bracketsWorker.findMaxLevelBrackets();
-        int maxBracketsLevel = bracketsWorker.getMaxBracketsLevel();
-        bracketsWorker.deleteAllBrackets();
-        System.out.println(bracketsWorker.toString());
+        try {
+            if (args.length != 1) {
+                throw new ArrayIndexOutOfBoundsException("You should enter only 1 parameter");
+            }
+            StringBuilder stringBuilder = new StringBuilder(args[0]);
+            BracketsWorker bracketsWorker = new BracketsWorker(stringBuilder);
+            bracketsWorker.deleteAllBrackets();
+            System.out.println(bracketsWorker.toString());
+        }
+        catch (ArrayIndexOutOfBoundsException ex){
+            System.out.println(ex.getMessage());
+        }
     }
 }
